service can noe parse and execute VOLUME actions

+++ IDo4U/app/src/main/java/com/my/ido4u/ActionDataClasses.kt
@@ -1,6 +1,6 @@
 package com.my.ido4u
 
-data class VolumeActionData (var volumeAction: VolumeAction, var volumeLevel: Float){
+data class VolumeActionData (var volumeAction: VolumeAction, var volumeLevel: Int){
     enum class VolumeAction {SOUND, MUTE, VIBRATE}
 }
